test: Run Typer commands in tests correctly. Move fixtures into common directory. Centralize most CLI tests.
